Removed all window functions from client
