+++ webAO/dom/addHPP.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Increment prosecution health point.
+ */
+export function addHPP() {
+    client.sendHP(2, client.hp[1] + 1);
+}
+window.addHPP = addHPP;
