+++ webAO/dom/guilty.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Declare the defendant not guilty
+ */
+export function guilty() {
+    client.sendRT("judgeruling#1");
+}
+window.guilty = guilty;
