+++ webAO/dom/initCE.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Declare cross examination.
+ */
+export function initCE() {
+    client.sendRT("testimony2");
+}
+window.initCE = initCE;
