+++ webAO/dom/initWT.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Declare witness testimony.
+ */
+export function initWT() {
+    client.sendRT("testimony1");
+}
+window.initWT = initWT;
