+++ webAO/dom/notGuilty.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Declare the defendant not guilty
+ */
+export function notguilty() {
+    client.sendRT("judgeruling#0");
+}
+window.notguilty = notguilty;
