+++ webAO/dom/randomCharacterOOC.ts
@@ -1,0 +1,8 @@
+import { client } from '../client'
+/**
+ * Random character via OOC.
+ */
+export function randomCharacterOOC() {
+    client.sendOOC(`/randomchar`);
+}
+window.randomCharacterOOC = randomCharacterOOC;
