+++ webAO/dom/redHPD.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Decrement defense health point.
+ */
+export function redHPD() {
+    client.sendHP(1, client.hp[0] - 1);
+}
+window.redHPD = redHPD;
