+++ webAO/dom/redHPP.ts
@@ -1,0 +1,9 @@
+import { client } from "../client";
+
+/**
+ * Decrement prosecution health point.
+ */
+export function redHPP() {
+    client.sendHP(2, client.hp[1] - 1);
+}
+window.redHPP = redHPP;
